fixed navbar rendering side by side issue
