+++ workout-tracker/src/History.js
@@ -1,0 +1,5 @@
+const NoPage = () => {
+    return <h1>404</h1>;
+  };
+  
+  export default NoPage;
