+++ workout-tracker/src/NavigationMenu.js
@@ -1,5 +1,5 @@
 import React from "react";
-
+import { Outlet, Link } from "react-router-dom";
 import'./css/NavigationMenu.css';
 // make css for nav bar
 // make the on clicks for the navbar buttons
@@ -8,13 +8,20 @@ import'./css/NavigationMenu.css';
 
 const Navbar = () => {
     return (
-    <div class="Navbar">
+    <div >
         
-         <button class="menuButton" type="button" value={"test1"}>Home</button>
-         <button class="menuButton" type="button" value={"test1"}>Create</button>
-         <button class="menuButton" type="button" value={"test1"}>History</button>
-         <button class="menuButton" type="button" value={"test1"}>Profile</button>
+        <div class="Navbar">
+            <Link to="/"><button class="menuButton" type="button" value={"test1"}>Home</button></Link>
+            <Link to="/create"><button class="menuButton" type="button" value={"test1"}>Create</button></Link>
          
+            <button class="menuButton" type="button" value={"test1"}>History</button>
+            <button class="menuButton" type="button" value={"test1"}>Profile</button>
+            <Link to="/nopage"><button class="menuButton" type="button" value={"test1"}>nopage</button></Link>
+        </div>
+         
+
+         <Outlet/>
+
     </div>
     )
 }
