+++ workout-tracker/src/css/NavigationMenu.css
@@ -4,16 +4,15 @@
     width: 100%;
     
     height: 100%;
-    border: 100px;
     
-    border-color: #ad0000;
+    position: relative;
     color: #000000;
     align-items: center;
     justify-content: center;
     
 }
 .menuButton {
-    display: inline;
+    display: block;
     min-width: fit-content;
     min-height: fit-content;
     width: 10%;
